Move some tests out of `electron-browser` (#175336)
* move some tests

* move more

* back to sandbox
